pyproject.toml
[project]
name = "AI-Sales-Agents"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "asyncio",
    "openai-agents>=0.0.15",
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,ollama,openai]>=0.4.9.2",
]
